+++ NexmoClient/Internal/NXMMemberPrivate.h
@@ -12,14 +12,14 @@
 
 @interface NXMMember (PrivateParser)
 
-- (nullable instancetype)initWithData:(NSDictionary *)data
+- (instancetype)initWithData:(NSDictionary *)data
                  andMemberIdFieldName:(NSString *)memberIdFieldName;
 
-- (nullable instancetype)initWithData:(NSDictionary *)data
+- (instancetype)initWithData:(NSDictionary *)data
                  andMemberIdFieldName:(NSString *)memberIdFieldName
                     andConversationId:(NSString *)convertaionId;
 
-- (nullable instancetype)initWithMemberEvent:(NXMMemberEvent *)memberEvent;
+- (instancetype)initWithMemberEvent:(NXMMemberEvent *)memberEvent;
 
 - (void)updateChannelWithLeg:(NXMLeg *)leg;
 - (void)updateMedia:(NXMMediaSettings *)media;