+++ NexmoClient/Internal/NexmoCore/Models/Events/Member/NXMMemberEvent.m
@@ -39,7 +39,7 @@
         self.memberId = memberId;
         self.user =  [[NXMUser alloc] initWithData:data[@"user"]];
         self.state = state;
-        self.media = [[NXMMediaSettings alloc] initWithEnabled:(data[@"media"] != nil ? YES : NO) suspend:NO];
+        self.media = [[NXMMediaSettings alloc] initWithEnabled:(data[@"media"] != nil ? [data[@"media"][@"audio"] boolValue] : NO) suspend:NO];
         self.knockingId = data[@"channel"][@"knocking_id"];
         self.channel = [[NXMChannel alloc] initWithData:data[@"channel"] andConversationId:conversationId andMemberId:memberId];
     }