+++ NexmoClient/NXMClientDelegate.h
@@ -18,6 +18,6 @@
 
 @optional
 - (void)incomingCall:(nonnull NXMCall *)call;
-- (void)addedToConversation:(nonnull NXMConversation *)conversation;
+- (void)incomingConversation:(nonnull NXMConversation *)conversation;
 
 @end