+++ NexmoClientTests/TestingUtils/NXMTestingUtils+Members.h
@@ -8,6 +8,7 @@
 
 #import "NXMTestingUtils.h"
 #import "NXMMember.h"
+#import "NXMMemberEvent.h"
 
 @interface NXMTestingUtils (Members)
 
@@ -15,5 +16,13 @@
 + (NXMMember *)memberWithConversationId:(NSString *)conversationId andUserId:(NSString *)userId state:(NXMMemberState)state;
 + (NXMMember *)memberWithConversationId:(NSString *)conversationId user:(NXMUser *)user state:(NXMMemberState)state;
 + (NXMMember *)memberWithConversationId:(NSString *)conversationId user:(NXMUser *)user state:(NXMMemberState)state memberId:(NSString *)memberId;
+
+
++ (NXMMemberEvent *)memberEventWithConvId:(NSString *)convId
+                                     user:(NSString *)userId
+                                    state:(NSString *)state
+                                 memberId:(NSString *)memberId
+                             fromMemberId:(NSString *)fromMemberId
+                                    media:(BOOL)media;
 
 @end