+++ NexmoClientTests/TestingUtils/NXMTestingUtils+Members.m
@@ -9,6 +9,7 @@
 #import "NXMTestingUtils+Members.h"
 #import "NXMUserPrivate.h"
 #import "NXMMemberPrivate.h"
+#import "NXMMemberEventPrivate.h"
 
 @implementation NXMTestingUtils (Members)
 
@@ -36,4 +37,55 @@
     return [[NXMMember alloc] initWithData:dict andMemberIdFieldName:@"memberId" andConversationId:conversationId];
 }
 
++ (NXMMemberEvent *)memberEventWithConvId:(NSString *)convId
+                                     user:(NSString *)userId
+                                    state:(NSString *)state
+                                 memberId:(NSString *)memberId
+                             fromMemberId:(NSString *)fromMemberId
+                                    media:(BOOL)media {
+    
+    NSDictionary *data = @{
+                        @"user": @{
+                            @"id": userId,
+                            @"user_id": userId,
+                            @"name": userId
+                        },
+                        @"channel": @{
+                            @"type": @"app",
+                            @"legs": @[],
+                            @"leg_settings": @{}
+                        },
+                        @"timestamp": @{
+                            state: @"2019-06-12T15:36:49.227Z"
+                        },
+                        @"initiator": @{
+                            state: @{
+                                @"isSystem": @(false),
+                                @"user_id": fromMemberId,
+                                @"member_id": fromMemberId
+                            }
+                        },
+                        @"media": @{
+                            @"audio_settings": @{
+                                @"enabled": @(media),
+                                @"earmuffed": @(false),
+                                @"muted": @(false)
+                            },
+                            @"audio": @(media)
+                        },
+                    };
+        
+    NXMMemberEvent *event = [[NXMMemberEvent alloc] initWithConversationId:@"1"
+                                                                sequenceId:1
+                                                                  andState:[state isEqualToString:@"invited"] ? NXMMemberStateInvited :
+                             [state isEqualToString:@"joined"] ? NXMMemberStateJoined :
+                             NXMMemberStateLeft
+                                                                   andData:data
+                                                              creationDate:nil
+                                                                  memberId:memberId];
+    
+    return event;
+    
+}
+    
 @end